pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "akgit"
description = "Akretion Git Custom"
authors = [{name = "Akretion", email = "contact@akretion.com"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]
readme = "README.rst"
dependencies = [
    "git-autoshare",
    "gitPython",
]
requires-python = ">=3.7"
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
    "coverage[toml]",
]

[project.urls]
Homepage = "https://github.com/akretion/akgit"
Source = "https://github.com/akretion/akgit"

[project.scripts]
supergit = "akgit.main:main"
sgit = "akgit.main:main"
git = "akgit.main:main"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build]
exclude = [
    "/.github",
]


[tool.ruff]
target-version = "py37"
fix = true

[tool.ruff.lint]
select = [
    "B",
    "E",
    "F",
    "I",
    "UP",
    "W",
]
ignore = [
]
exclude = [
    "docs/conf.py",
]

[tool.ruff.lint.isort]
known-first-party = []


[tool.coverage.run]
branch = true
source_pkgs = ["akgit"]

[tool.coverage.paths]
source = ["src", ".tox/*/site-packages"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
]



[tool.towncrier]
package = "akgit"
package_dir = "src"
filename = "HISTORY.rst"
directory = "news"
issue_format = "`#{issue} <https://github.com/akretion/akgit/issues/{issue}>`_"
title_format = "{version} ({project_date})"
